chore: 커뮤니티 글 전체 조회 API 수정
- page, size 타입 변경 (int -> long)
- 불필요한 출력문 제거

+++ src/main/java/com/zeroway/community/repository/post/PostRepositoryCustom.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public interface PostRepositoryCustom {
 
-    List<PostListRes> getPostList(Long userId, String sort, Boolean challenge, Boolean review, int page, int size);
+    List<PostListRes> getPostList(Long userId, String sort, Boolean challenge, Boolean review, long page, long size);
     PostRes getPost(Long postId, Long userId);
     List<UserInfo> getPostLikeList(Long postId);
 }
